pyproject.toml
[tool.poetry]
name = "delta-wrapper"
version = "0.1.0"
description = ""
authors = ["Jens Peder Meldgaard <jenspederm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyspark = "^3.3.1"
delta-spark = "^2.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
